requirements.txt
earthengine-api
geemap
setuptools
plotly
pandas
imageio
requests
pillow
numpy